+++ test/crdt_test.dart
@@ -13,28 +13,27 @@ void main() {
 
     test('Put', () async {
       await crdt.put('x', 1);
-      var record = await crdt.get('x');
-      expect(record.value, 1);
+      var value = crdt.get('x');
+      expect(value, 1);
     });
 
     test('Put sequential', () async {
       await crdt.put('x', 1);
       await crdt.put('x', 2);
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Put many', () async {
       await crdt.putAll({'x': 2, 'y': 3});
-      expect((await crdt.get('x')).value, 2);
-      expect((await crdt.get('y')).value, 3);
+      expect(crdt.get('x'), 2);
+      expect(crdt.get('y'), 3);
     });
 
     test('Delete value', () async {
       await crdt.put('x', 1);
       await crdt.delete('x');
-      var record = await crdt.get('x');
-      expect(record.isDeleted, isTrue);
+      expect(crdt.isDeleted('x'), isTrue);
     });
   });
 
@@ -45,15 +44,15 @@ void main() {
       crdt = Crdt(MapStore({'x': Record(Hlc(), 1)}));
     });
 
-    test('Seed item', () async {
-      var record = await crdt.get('x');
-      expect(record.value, 1);
+    test('Seed item', () {
+      var value = crdt.get('x');
+      expect(value, 1);
     });
 
     test('Seed and put', () async {
       await crdt.put('x', 2);
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
   });
 
@@ -68,58 +67,57 @@ void main() {
     test('Merge older', () async {
       await crdt.put('x', 2);
       await crdt.merge({'x': Record(Hlc(now - 10000), 1)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge very old', () async {
       await crdt.put('x', 2);
       await crdt.merge({'x': Record(Hlc(now - 1000000), 1)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge newer', () async {
       await crdt.put('x', 1);
       await crdt.merge({'x': Record(Hlc(now + 1000000), 2)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge same', () async {
       await crdt.put('x', 2);
-      var remoteTs = (await crdt.get('x')).hlc;
+      var remoteTs = crdt.getRecord('x').hlc;
       await crdt.merge({'x': Record(remoteTs, 1)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge older, newer counter', () async {
       await crdt.put('x', 2);
       await crdt.merge({'x': Record(Hlc(now - 1000000, 2), 1)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge same, newer counter', () async {
       await crdt.put('x', 1);
-      var remoteTs = Hlc((await crdt.get('x')).hlc.micros, 2);
+      var remoteTs = Hlc(crdt.getRecord('x').hlc.micros, 2);
       await crdt.merge({'x': Record(remoteTs, 2)});
-      var record = await crdt.get('x');
-      expect(record.value, 2);
+      var value = crdt.get('x');
+      expect(value, 2);
     });
 
     test('Merge new item', () async {
       var map = {'x': Record<int>(Hlc(), 2)};
       await crdt.merge(map);
-      expect(await crdt.getMap(), map);
+      expect(crdt.getMap(), map);
     });
 
     test('Merge deleted item', () async {
       await crdt.put('x', 1);
       await crdt.merge({'x': Record(Hlc(now + 1000000), null)});
-      var record = await crdt.get('x');
-      expect(record.isDeleted, isTrue);
+      expect(crdt.isDeleted('x'), isTrue);
     });
   });
 
@@ -130,27 +128,27 @@ void main() {
       crdt = Crdt(MapStore({'x': Record<int>(Hlc(1579633503110), 1)}));
     });
 
-    test('To map', () async {
-      expect(await crdt.getMap(), {'x': Record<int>(Hlc(1579633503110), 1)});
+    test('To map', () {
+      expect(crdt.getMap(), {'x': Record<int>(Hlc(1579633503110), 1)});
     });
 
-    test('jsonEncodeStringKey', () async {
-      expect(jsonEncode(await crdt.getMap()),
+    test('jsonEncodeStringKey', () {
+      expect(jsonEncode(crdt.getMap()),
           '{"x":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":1}}');
     });
 
-    test('jsonEncodeIntKey', () async {
+    test('jsonEncodeIntKey', () {
       expect(crdtMap2Json({1: Record(Hlc.fromLogicalTime(1579633475584), 1)}),
           '{"1":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":1}}');
     });
 
-    test('jsonDecodeStringKey', () async {
+    test('jsonDecodeStringKey', () {
       var map = json2CrdtMap<String, int>(
           '{"x":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":1}}');
-      expect(map, await crdt.getMap());
+      expect(map, crdt.getMap());
     });
 
-    test('jsonDecodeIntKey', () async {
+    test('jsonDecodeIntKey', () {
       var map = json2CrdtMap<int, int>(
           '{"1":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":1}}',
           keyDecoder: (key) => int.parse(key));
@@ -166,21 +164,21 @@ void main() {
           {'x': Record<TestClass>(Hlc(1579633503110), TestClass('test'))}));
     });
 
-    test('To map', () async {
-      expect(await crdt.getMap(),
+    test('To map', () {
+      expect(crdt.getMap(),
           {'x': Record<TestClass>(Hlc(1579633503110), TestClass('test'))});
     });
 
-    test('jsonEncode', () async {
-      expect(jsonEncode(await crdt.getMap()),
+    test('jsonEncode', () {
+      expect(jsonEncode(crdt.getMap()),
           '{"x":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":{"test":"test"}}}');
     });
 
-    test('jsonDecode', () async {
+    test('jsonDecode', () {
       var decoded = json2CrdtMap<String, TestClass>(
           '{"x":{"hlc":"1970-01-19T06:47:13.476Z-0000","value":{"test":"test"}}}',
           valueDecoder: TestClass.fromJson);
-      expect(await decoded, await crdt.getMap());
+      expect(decoded, crdt.getMap());
     });
   });
 }